Detected where the zoom bug comes from, and changed some stuff

+++ src/main/java/fr/gallon_labergere/fractales/model/Settings.java
@@ -95,7 +95,7 @@ public class Settings extends Observable {
      * @param x
      * @return
      */
-    public float getViewX(int x) {
+    public float getViewX(float x) {
         return ((x - centerX) / zoomLevel);
     }
 
@@ -104,7 +104,7 @@ public class Settings extends Observable {
      * @param y
      * @return
      */
-    public float getViewY(int y) {
+    public float getViewY(float y) {
         return ((y - centerY) / zoomLevel);
     }
